날씨 테스트

+++ src/main/java/com/bot/gachon/dto/res/WeatherSub2Dto.java
@@ -8,14 +8,14 @@ import lombok.Getter;
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Getter
 public class WeatherSub2Dto {
-    private Double temp;
-    private Double temp_min;
-    private Double temp_max;
+    private Long temp;
+    private Long temp_min;
+    private Long temp_max;
     private Long humidity;
 
     public WeatherSub2Dto(){}
     @Builder
-    public WeatherSub2Dto(Double temp, Double temp_max, Double temp_min,Long humidity){
+    public WeatherSub2Dto(Long temp, Long temp_max, Long temp_min,Long humidity){
         this.temp = temp;
         this.temp_max = temp_max;
         this.temp_min = temp_min;
